requirements.txt
pip
setuptools
wheel
coverage
sphinx
recommonmark
sphinx_rtd_theme

GitPython
lxml
nltk
beautifulsoup4
python-slugify